Saving the 'index' from previous STATE
Problem:
RemoteBackend uses 'index' value for replying to queries from the
'rotation' list. After every response is added , 'index' was incremented
but was not committed to memcached.
So mostly this 'index' would remain unchanged in the memcached as its never
set into memcached.
For every fetch 'get' from polaris_pdns.StateUpdater() it would start from 0
and this leads to responses not being sent according to Round robin algo.

Solution:
Saving the previous 'index' in dist_table every time in StateUpdater()
so that responses would be actually distributed according to RR algo.